fix: move UUID PK serialization into pgai lib
In b242d7049a1c38785c510f0a5a36af31537cb610 we introduced a fix for
broken UUID serialization. Unfortunately this fix doesn't apply to
consumers of pgai as a library.

This change moves the fix from the cli to the upper-most level in the
worker.
